Send unassigned event when deleting tag.

+++ Swipes/swipes/src/main/java/com/swipesapp/android/sync/service/TasksService.java
@@ -4,6 +4,10 @@ import android.content.Context;
 import android.content.Intent;
 
 import com.swipesapp.android.analytics.handler.Analytics;
+import com.swipesapp.android.analytics.handler.IntercomHandler;
+import com.swipesapp.android.analytics.values.IntercomEvents;
+import com.swipesapp.android.analytics.values.IntercomFields;
+import com.swipesapp.android.analytics.values.Labels;
 import com.swipesapp.android.app.SwipesApplication;
 import com.swipesapp.android.db.Attachment;
 import com.swipesapp.android.db.DaoSession;
@@ -25,6 +29,7 @@ import java.lang.ref.WeakReference;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Date;
+import java.util.HashMap;
 import java.util.List;
 
 /**
@@ -364,8 +369,10 @@ public class TasksService {
     public void deleteTag(Long tagId) {
         // Unassign from all tasks.
         List<TaskTag> assignments = mExtTaskTagDao.selectAssociationsForTag(tagId);
-        for (TaskTag assignment : assignments) {
-            mExtTaskTagDao.getDao().delete(assignment);
+        if (assignments != null) {
+            for (TaskTag assignment : assignments) {
+                mExtTaskTagDao.getDao().delete(assignment);
+            }
         }
 
         // Delete from database.
@@ -376,6 +383,9 @@ public class TasksService {
 
         // Update number of tags dimension.
         Analytics.sendNumberOfTags(mContext.get());
+
+        // Send unassign event.
+        sendTagUnassignEvent(assignments);
     }
 
     /**
@@ -1000,6 +1010,23 @@ public class TasksService {
         }
 
         return attachments;
+    }
+
+    /**
+     * Sends a tag unassigned event.
+     *
+     * @param assignments Tasks this tag is being unassigned from.
+     */
+    private void sendTagUnassignEvent(List assignments) {
+        // Prepare Intercom fields.
+        HashMap<String, Object> fields = new HashMap<>();
+        fields.put(IntercomFields.NUMBER_OF_TAGS, 1);
+        fields.put(IntercomFields.FROM, Labels.TAGS_FROM_DELETE_TAG);
+
+        if (assignments != null) fields.put(IntercomFields.NUMBER_OF_TASKS, assignments.size());
+
+        // Send Intercom events.
+        IntercomHandler.sendEvent(IntercomEvents.UNASSIGN_TAGS, fields);
     }
 
 }
